mtp: adding support for merging folders
Returning ERR_DIR_ALREADY_EXIST for mkdir will pop up a merge window and continue based on users selection.

+++ mtp/kiod_module/mtpstorage.h
@@ -129,7 +129,7 @@ public Q_SLOTS:
     int setFileName(const QString &path, const QString &newName);
 
     // folder management
-    quint32 createFolder(const QString &path);
+    qint64 createFolder(const QString &path);
 
     // object management
     int deleteObject(const QString &path);
